Added SmsDto class

+++ src/main/java/com/meesho/mohsin/NotificationService/repository/SmsRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface SmsRepository extends CrudRepository<SmsRequests,Long> {
+public interface SmsRepository extends CrudRepository<SmsRequests,Integer> {
 
     List<SmsRequests> findAll();
-    SmsRequests findById(int id);  // why can't i use Long ?
+    SmsRequests findById(int id);
 
 }
